Removed the auto-adjust-for-fewer-windows behaviour.
This is because the way I've set up the events, it doesn't
re-place the entire page's windows when a new window appears
(which is a feature, not a bug) but that means that the previously
placed windows will be the wrong size.  So don't do the "smart"
adjusting, just assume that the user wants the layout they chose.